test(upstream): attach through the mounted repo, keep legacy endpoint coverage
