+++ vestad/tests/server/mod.rs
@@ -15,4 +15,4 @@ mod ports;
 mod agent_code;
 mod layout;
 mod rename;
-mod workspace;
+mod upstream;
